Move USB backends to separate file
