+++ dali/driver/tridonic.py
@@ -2,8 +2,8 @@ from dali.command import from_frame
 from dali.driver.base import AsyncDALIDriver
 from dali.driver.base import DALIDriver
 from dali.driver.base import SyncDALIDriver
-from dali.driver.base import USBBackend
-from dali.driver.base import USBListener
+from dali.driver.usb import USBBackend
+from dali.driver.usb import USBListener
 from dali.frame import BackwardFrame
 from dali.frame import ForwardFrame
 import logging
